+++ Sharp2048/Sharp2048/GameLogic.cs
@@ -1,6 +1,0 @@
-﻿namespace Sharp2048
-{
-    class GameLogic
-    {
-    }
-}