+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [v1.2.1] - (08.09.21)
+
+### Added
+
+ - Config file (still no menu);
+ - Icon;
+ - Rounded palletes;
+
+### Changed
+
+ - Clean code;
+ - Authentic colors;
+ 
+### Fixed
+
+ - Minor WM version fixes;
+
 ## [v1.2.0] - (31.08.21)
 
 ### Added
@@ -20,5 +37,6 @@
 
  - First release;
 
+[v1.2.1]: https://github.com/Limows/Sharp2048/releases/tag/v1.2.1
 [v1.2.0]: https://github.com/Limows/Sharp2048/releases/tag/v1.2.0
 [v1.0.0]: https://github.com/Limows/Sharp2048/releases/tag/v1.0.0